refactor(ui): lift pool state to ValidatorDetails
Move pool-related state and logic from `StakingDetails` to `ValidatorDetails`
component to avoid duplicate `AddPoolModal` instances. Extract pool selection
logic into `SelectPool` component.

- Delete `StakingDetails` component and move its logic to `ValidatorDetails`
- Extract pool selection logic into `SelectPool` component
- Move `AddPoolModal` state and `canAddPool` logic to `ValidatorDetails`
- Pass pool-related props from `ValidatorDetails` to `Highlights`
- Add `StakerChartData` interface
